refactor: based on call w Doris and Johno
